refactor: split logic and UI into two files
- logic to generate pix codes, set loading states, redirect user and
copy code to clipboard is at `usePixCode` (which can be refactored
further)
- UI is on `page.tsx`. `PixDonationPage.tsx` was deleted
